+++ accelerator/device/device_structures.hpp
@@ -88,7 +88,7 @@ typedef struct {
     signed char cluster_values [CLUSTER_SIZE];
 } t_cluster;
 
-typedef short t_bias;
+typedef signed int t_bias;
 #else
 // #if (ACCUMULATOR_WIDTH == 32)
 // typedef signed int t_accumulator;
@@ -104,7 +104,7 @@ typedef struct {
 
 typedef cl_ushort t_streamblock_address;
 
-typedef signed short t_bias;
+typedef signed int t_bias;
 #endif
